requirements.txt
google-cloud-dlp
google-cloud-bigquery
pandas
typing
